第一次联调完

+++ tropical-disease-research/src/main/java/team/t9001/saad/controller/ArticleController.java
@@ -8,11 +8,12 @@ import org.springframework.stereotype.Controller;
 import org.springframework.web.bind.annotation.RequestMapping;
 import org.springframework.web.bind.annotation.RequestMethod;
 import org.springframework.web.bind.annotation.ResponseBody;
-import org.springframework.web.servlet.ModelAndView;
 import team.t9001.saad.common.*;
 import team.t9001.saad.model.Article;
 import team.t9001.saad.service.ArticleService;
+import team.t9001.saad.service.ValidatorService;
 
+import javax.servlet.http.HttpServletRequest;
 import java.util.List;
 
 /**
@@ -26,6 +27,8 @@ public class ArticleController {
 
     @Autowired
     private ArticleService articleService;
+    @Autowired
+    private ValidatorService validatorService;
 
     /**
      * 添加文章
@@ -33,10 +36,40 @@ public class ArticleController {
      */
     @ResponseBody
     @RequestMapping(value = UrlConstants.add_article, method = RequestMethod.POST)
-    public int addArticle(Article article){
-        logger.info("add article begin.");
-        int result = articleService.addArticle(article);
-        return result;
+    public RequestStatus addArticle(HttpServletRequest request, Article article){
+        logger.debug("add article begin, article: {}", article);
+        RequestStatus requestStatus = new RequestStatus();
+
+        //校验参数
+        if (article == null
+                || article.getTitle() == null
+                || article.getExternalLink() == null
+                || article.getStatus() < 0) {
+            requestStatus.setStatus(Constants.STATUS_NOT_OK);
+            requestStatus.setErrorCode(ErrorCode.INPUT_ERROR);
+            requestStatus.setErrorMsg(ErrorCode.ERROR_MSG_MAP.get(ErrorCode.INPUT_ERROR) + "，article = " + article);
+            return requestStatus;
+        }
+
+        //校验权限
+        if (!validatorService.validateAdminPermission(requestStatus, request)){
+            return requestStatus;
+        }
+
+        try {
+            int result = articleService.addArticle(article);
+            if (result != 1) {
+                requestStatus.setStatus(Constants.STATUS_NOT_OK);
+                requestStatus.setErrorCode(ErrorCode.NO_ROW_AFFECTED);
+                requestStatus.setErrorMsg(ErrorCode.ERROR_MSG_MAP.get(ErrorCode.NO_ROW_AFFECTED));
+            }
+            logger.info("add article success, article:{}.", article);
+        } catch (Exception e) {
+            requestStatus.setStatus(Constants.STATUS_NOT_OK);
+            requestStatus.setErrorCode(ErrorCode.DB_ACCESS_ERROR);
+            requestStatus.setErrorMsg(ErrorCode.ERROR_MSG_MAP.get(ErrorCode.DB_ACCESS_ERROR));
+        }
+        return requestStatus;
     }
 
     /**
@@ -44,24 +77,23 @@ public class ArticleController {
      * @param page
      * @return
      */
+    @ResponseBody
     @RequestMapping(value = UrlConstants.get_article_list, method = RequestMethod.GET)
-    public ModelAndView getArticleList(Page page, Integer userId){
-        ModelAndView mv = new ModelAndView("article/list");
-
+    public RequestStatus getArticleList(HttpServletRequest request, Page page){
         RequestStatus requestStatus = new RequestStatus();
-        if (!Validator.validateAdminRights(requestStatus, userId)
-                || !Validator.validatePage(requestStatus, page)) {
-            mv.addObject(requestStatus);
-            return mv;
+        if (!validatorService.validateLogin(requestStatus, request)) {
+            return requestStatus;
+        }
+
+        if (!validatorService.validatePage(requestStatus, page)) {
+            return requestStatus;
         }
 
         List<Article> articleList = articleService.getArticleList(page);
-        logger.info("get article list, page:{}, data:{}", JSON.toJSONString(page), JSON.toJSONString(articleList));
+        logger.info("get article list, page:{}, data:{}", page, articleList);
 
-
-        mv.addObject("list", articleList);
-        mv.addObject("requestStatus", requestStatus);
-        return mv;
+        requestStatus.setData(JSON.toJSONString(articleList));
+        return requestStatus;
     }
 
     /**
@@ -71,9 +103,41 @@ public class ArticleController {
      */
     @ResponseBody
     @RequestMapping(value = UrlConstants.modify_article, method = RequestMethod.POST)
-    public int modifyArticle(Article article){
-        int result = articleService.modifyArticle(article);
-        return result;
+    public RequestStatus modifyArticle(HttpServletRequest request, Article article){
+        logger.debug("modify article begin.");
+        RequestStatus requestStatus = new RequestStatus();
+
+        //校验参数
+        if (article == null
+                || article.getArticleId() <= 0
+                || article.getTitle() == null
+                || article.getExternalLink() == null
+                || article.getStatus() < 0) {
+            requestStatus.setStatus(Constants.STATUS_NOT_OK);
+            requestStatus.setErrorCode(ErrorCode.INPUT_ERROR);
+            requestStatus.setErrorMsg(ErrorCode.ERROR_MSG_MAP.get(ErrorCode.INPUT_ERROR) + "，article = " + article);
+            return requestStatus;
+        }
+
+        //校验权限
+        if (!validatorService.validateAdminPermission(requestStatus, request)){
+            return requestStatus;
+        }
+
+        try {
+            int result = articleService.modifyArticle(article);
+            if (result != 1) {
+                requestStatus.setStatus(Constants.STATUS_NOT_OK);
+                requestStatus.setErrorCode(ErrorCode.NO_ROW_AFFECTED);
+                requestStatus.setErrorMsg(ErrorCode.ERROR_MSG_MAP.get(ErrorCode.NO_ROW_AFFECTED));
+            }
+            logger.info("modify article success, article:{}.", article);
+        } catch (Exception e) {
+            requestStatus.setStatus(Constants.STATUS_NOT_OK);
+            requestStatus.setErrorCode(ErrorCode.DB_ACCESS_ERROR);
+            requestStatus.setErrorMsg(ErrorCode.ERROR_MSG_MAP.get(ErrorCode.DB_ACCESS_ERROR));
+        }
+        return requestStatus;
     }
 
     /**
@@ -83,8 +147,35 @@ public class ArticleController {
      */
     @ResponseBody
     @RequestMapping(value = UrlConstants.remove_article, method = RequestMethod.POST)
-    public int removeArticle(Integer articleId){
-        int result = articleService.removeArticle(articleId);
-        return result;
+    public RequestStatus removeArticle(HttpServletRequest request, Integer articleId){
+        RequestStatus requestStatus = new RequestStatus();
+
+        //校验参数
+        if (articleId == null) {
+            requestStatus.setStatus(Constants.STATUS_NOT_OK);
+            requestStatus.setErrorCode(ErrorCode.INPUT_ERROR);
+            requestStatus.setErrorMsg(ErrorCode.ERROR_MSG_MAP.get(ErrorCode.INPUT_ERROR) + "，articleId = null");
+            return requestStatus;
+        }
+
+        //校验权限
+        if (!validatorService.validateAdminPermission(requestStatus, request)){
+            return requestStatus;
+        }
+
+        try {
+            int result = articleService.removeArticle(articleId);
+            if (result != 1) {
+                requestStatus.setStatus(Constants.STATUS_NOT_OK);
+                requestStatus.setErrorCode(ErrorCode.NO_ROW_AFFECTED);
+                requestStatus.setErrorMsg(ErrorCode.ERROR_MSG_MAP.get(ErrorCode.NO_ROW_AFFECTED));
+            }
+            logger.info("remove article success, articleId:{}.", articleId);
+        } catch (Exception e) {
+            requestStatus.setStatus(Constants.STATUS_NOT_OK);
+            requestStatus.setErrorCode(ErrorCode.DB_ACCESS_ERROR);
+            requestStatus.setErrorMsg(ErrorCode.ERROR_MSG_MAP.get(ErrorCode.DB_ACCESS_ERROR));
+        }
+        return requestStatus;
     }
 }
